+++ src/pages/Main/_components/VotingItem/FullVotingItem.tsx
@@ -5,15 +5,14 @@ import Budget from "../../../../assets/icons/budget.svg";
 import Schedule from "../../../../assets/icons/calendar3.svg";
 import calculateVoteGauge from "./calculateVoteGauge";
 import renderParticipantProfiles from "./renderParticipantProfiles";
-import useRemainingTimes from "./useRemainingTimes";
+import RemainingTimeDisplay from "./useRemainingTimes";
 import MapComponent from "../../../SearchPage/_components/SearchMap/SearchMap";
 import { useNavigate } from "react-router-dom";
-import { memo } from "react";
 
-interface RemainingTimeProps {
-    tripPlanId: number;
-    remainingTimes: { [key: number]: string };
-}
+// interface RemainingTimeProps {
+//     tripPlanId: number;ｚＺＸＸｚＸｘｚＸｚＸ？！ｄｊｓａ
+//     remainingTimes: { [key: number]: string };
+// }
 
 interface FullVotingItemProps {
     rooms: Room[];
@@ -36,15 +35,16 @@ interface Room {
     remainingTime: "THIRTY_MINUTES" | "SIXTY_MINUTES" | "FOUR_HOURS" | "SIX_HOURS";
 }
 
-// 타이머 부분만 메모이제이션
-const RemainingTimeDisplay: React.FC<RemainingTimeProps> = memo(({ tripPlanId, remainingTimes }) => {
-    return <span>{remainingTimes[tripPlanId] || "00:00:00"}</span>;
-});
+// // 타이머 부분만 메모이제이션
+// const RemainingTimeDisplay: React.FC<RemainingTimeProps> = memo(({ tripPlanId, remainingTimes }) => {
+//     return <span>{remainingTimes[tripPlanId] || "00:00:00"}</span>;
+// });
 
 const FullVotingItem: React.FC<FullVotingItemProps> = ({ rooms = [], selectedFilter }) => {
-    const remainingTimes = useRemainingTimes(rooms);
+    // const remainingTimes = useRemainingTimes(rooms);
     const navigate = useNavigate();
 
+    console.log("렌더링중");
     const handleClick = (tripPlanId: number, roomId: string, tripPlanType: "COURSE" | "SCHEDULE" | "BUDGET", masterId: string) => {
         if (tripPlanType === "COURSE") {
             // COURSE 타입일 때
@@ -84,11 +84,12 @@ const FullVotingItem: React.FC<FullVotingItemProps> = ({ rooms = [], selectedFil
                         <V.Box>
                             <S.Box>
                                 <V.Title>{room.roomName}</V.Title>
-                                {/* 타이머 부분만 리렌더링 */}
+                                {/* 🔥 개별 타이머 컴포넌트로 리렌더링 최소화 */}
                                 <V.RemainingTime>
-                                    <RemainingTimeDisplay tripPlanId={room.tripPlanId} remainingTimes={remainingTimes} />
+                                    <RemainingTimeDisplay createdAt={room.createdAt} remainingTime={room.remainingTime} />
                                 </V.RemainingTime>
                             </S.Box>
+
                             <S.Box>
                                 <S.Location>{room.location}</S.Location>
                             </S.Box>
